Change getPossibilitiesCount method return type to long. Some implementations may return values too big to store in int.

+++ src/main/java/pl/dzielins42/dmtools/generator/GeneratorStatistics.java
@@ -4,5 +4,5 @@ public interface GeneratorStatistics {
     
     public static final int UNKNOWN = -1;
     
-    int getPossibilitiesCount();
+    long getPossibilitiesCount();
 }
