+++ src/main/java/pl/dzielins42/dmtools/generator/name/BasicNameGenerator.java
@@ -20,7 +20,7 @@ public class BasicNameGenerator implements NameGenerator {
     }
 
     @Override
-    public int getPossibilitiesCount() {
+    public long getPossibilitiesCount() {
         return names.length;
     }
